Programmatically handle RMI Registry.
Avoids the non-sense of having to muck around with the rmiregistry
executable since all of that stuff is hidden from inside the code anyway.

I've exposed the ability to launch on a non-default port, but you should
be leery about it. Sciore's original form has localhost:1099 assumptions
baked into it in multiple places. The launcher scripts use the default
port for RMI Registries.